requirements.txt
Flask
sentence-transformers
PyPDF2
python-docx
torch
gunicorn 
numpy
pytest